% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{dist_plot}
\alias{dist_plot}
\title{Bayesian Equivalence Testing (BET): Visualize Model}
\usage{
dist_plot(df, main, lower, upper)
}
\arguments{
\item{main}{Title of Graph}

\item{lower}{Upper bound of x axis}

\item{model_name}{Name of model}
}
\description{
This function reduces six important BET inspection steps into one. That is, this helper function
cautomates the following: 1) visualizing your model (relies on 'sjPlot');
2) examine priors (relies on)
}
\examples{
tess_glm <- stan_glm(FINAL_1mother_0other ~ condition, family = "binomial",
data = tess_analyze[tess_analyze$seen_before == 3, ], iter = 10000, seed = 1839,
diagnostic_file = file.path(tempdir(), "df.csv"))
 inspect_model(tess_glm)
}
\keyword{Bayesian}
\keyword{Equivalence}
\keyword{Testing}
