% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_model.R
\name{inspect_model}
\alias{inspect_model}
\title{Bayesian Equivalence Testing (BET): Inspect Model}
\usage{
inspect_model(model_name)
}
\arguments{
\item{model_name}{Name of model}
}
\description{
This function reduces six important BET inspection steps into one. That is, this helper function
cautomates the following: 1) visualizing your model (relies on 'sjPlot');
2) examine priors (relies on)
}
\examples{
tess_glm <- stan_glm(FINAL_1mother_0other ~ condition, family = "binomial",
data = tess_analyze[tess_analyze$seen_before == 3, ], iter = 10000, seed = 1839,
diagnostic_file = file.path(tempdir(), "df.csv"))
 inspect_model(tess_glm)
}
\keyword{Bayesian}
\keyword{Equivalence}
\keyword{Testing}
